backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic>=2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
azure-storage-blob==12.19.0
aiohttp>=3.8.0 